Update sessions helper
