+++ app/helpers/sessions_helper.rb
@@ -1,0 +1,14 @@
+module SessionsHelper
+  
+  def login(user)
+    session[:user_id] = user.id
+  end
+
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id])
+  end
+
+  def logged_in?
+    !current_user.nil?
+  end
+end
