+++ app/views/layouts/application.html.erb
@@ -16,7 +16,11 @@
         <ul class="main-nav">
           <li> <%= link_to('Home', root_path)%> </li>
           <li> <%= link_to('Sign Up', new_user_path)%> </li>
-          <li> <%= link_to('Login', login_path)%> </li>
+          <% if logged_in? %>
+            <li> <%= link_to('Logout', logout_path)%> </li>
+          <% else %>
+            <li> <%= link_to('Login', login_path)%> </li>
+          <% end %>
         </ul>
       </nav>
     </div>
