+++ app/views/users/new.html.erb
@@ -6,7 +6,7 @@
     </div>
   </div>
   <div class="row">
-    <div class="col-sm-6 col-md-6 form-container"> 
+    <div class="col-sm-6 col-md-6 col-lg-6 form-container"> 
       <%= form_for(@user) do |f| %>
         <%= render "shared/error_messages" %>
         <div class="form-group">
